% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invMGF.R
\name{invMGF}
\alias{invMGF}
\title{Compute the Moments of IV}
\usage{
invMGF(m, gamma, v0, v1, tau, k, theta, sigma)
}
\arguments{
\item{m}{largest order of the moments to be computed.}

\item{gamma}{accuracy parameter, error bound \eqn{10^{-\gamma}}.}

\item{v0}{the left end, \eqn{v_u}.}

\item{v1}{the right end, \eqn{v_t}.}

\item{tau}{time difference, \eqn{t-u}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}
}
\value{
vector of moments, \eqn{\mu_1,\cdots,\mu_m}.
}
\description{
Compute the moments of IV by numerical inversion of an adaptively modified
moment generating function introduced by Choudhury and Lucantoni (1996).
}
\examples{
k = 6.21; theta = 0.019; sigma = 0.61; v0 = 0.010201; v1 = v0; tau = 1
m = 4; gamma = 11
mu = invMGF(m, gamma, v0, v1, tau, k, theta, sigma)
}
