% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplace.R
\name{Laplace}
\alias{Laplace}
\title{Laplace Transform of IV}
\usage{
Laplace(a, v0, v1, tau, k, theta, sigma)
}
\arguments{
\item{a}{Laplace transform input, can be either a real number or a complex
number.}

\item{v0}{the left end, \eqn{v_u}.}

\item{v1}{the right end, \eqn{v_t}.}

\item{tau}{time difference, \eqn{t-u}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}
}
\value{
Laplace transform result, either a scalar real number or a scalar
complex number, corresponding to the input whether a real number or a
complex number.
}
\description{
Laplace transform of Integrated Variance (IV) with condition on two side
levels \eqn{v_u} and \eqn{v_t}. It is different from Moment-Generating
function:
\eqn{
  \mathcal{L}(a) = M(-a).
}
Closely related with the Characteristic Function through
\eqn{\Phi(a) = \mathcal{L}(-ia)}, where \eqn{\Phi(a)} is the Characteristic
Function which is implemented as the function \code{\link[ajd.sim.bk:CF]{ajd.sim.bk::CF()}}. The
involved BesselI function is evaluated by \code{\link[Bessel:Bessel]{Bessel::BesselI()}} or
\code{\link[Bessel:besselIasym]{Bessel::besselIasym()}}.
}
\examples{
k = 6.21; theta = 0.019; sigma = 0.61; v0 = 0.010201; v1 = v0; tau = 1
a = 1
Laplace(a, v0, v1, tau, k, theta, sigma)
}
\seealso{
Closely related with the Characteristic Function \code{\link[ajd.sim.bk:CF]{ajd.sim.bk::CF()}}.
}
