% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_svcj.R
\name{price_svcj}
\alias{price_svcj}
\title{Pricing the European Call Option Under the SVCJ Model}
\usage{
price_svcj(
  N,
  S,
  K,
  v0,
  tau,
  r,
  k,
  theta,
  sigma,
  rho,
  lambda,
  mu_bar,
  sigma_s,
  mu_v,
  rho_J,
  true_price
)
}
\arguments{
\item{N}{number of underling asset price samples to simulate.}

\item{S}{current underling asset price.}

\item{K}{striking price.}

\item{v0}{current variance level.}

\item{tau}{maturity time.}

\item{r}{riskless rate.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}

\item{rho}{parameter \eqn{\rho}.}

\item{lambda}{parameter \eqn{\lambda}, arrival rate of jumps in the
underling asset price process and variance process.}

\item{mu_bar}{parameter \eqn{\bar{\mu}}, see Broadie-Kaya (2006).}

\item{sigma_s}{parameter \eqn{\sigma_s}, see Broadie-Kaya (2006).}

\item{mu_v}{parameter \eqn{\mu_v}, mean of the jumps
(exponential distribution) in the variance.}

\item{rho_J}{parameter \eqn{\rho_J}, correlation parameter between the
jumps in the price and variance.}

\item{true_price}{theoretical true price of the option, calculable by
the Duffie et al. (2000) method.}
}
\value{
vector of pricing error and consumed time (simulation).
}
\description{
Pricing the European call option under the SVCJ model using the
Kyriakou-Brignone-Fusai (2024) method for the simulation.
}
\examples{
S = 100; K = 100; v0 = 0.007569; k = 3.46; theta = 0.008; sigma = 0.14
rho = -0.82; r = 0.0319; tau = 1; lambda = 0.47; mu_bar = -0.1
sigma_s = 0.0001; mu_v = 0.05; rho_J = -0.38; true_price = 6.8619
# price_svcj(10000, S, K, v0, tau, r, k, theta, sigma, rho, lambda, mu_bar,
# sigma_s, mu_v, rho_J, true_price)
}
