% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_svj.R
\name{price_svj}
\alias{price_svj}
\title{Pricing the European Call Option Under the SVJ Model}
\usage{
price_svj(
  N,
  S,
  K,
  v0,
  tau,
  r,
  k,
  theta,
  sigma,
  rho,
  lambda,
  mu_bar,
  sigma_s,
  true_price
)
}
\arguments{
\item{N}{number of underling asset price samples to simulate.}

\item{S}{current underling asset price.}

\item{K}{striking price.}

\item{v0}{current variance level.}

\item{tau}{maturity time.}

\item{r}{riskless rate.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}

\item{rho}{parameter \eqn{\rho}.}

\item{lambda}{parameter \eqn{\lambda}, arrival rate of jumps in the
underling asset price process.}

\item{mu_bar}{parameter \eqn{\bar{\mu}}, see Broadie-Kaya (2006).}

\item{sigma_s}{parameter \eqn{\sigma_s}, see Broadie-Kaya (2006).}

\item{true_price}{theoretical true price of the option, calculable by
the Bates (1996) method.}
}
\value{
vector of pricing error and consumed time (simulation).
}
\description{
Pricing the European call option under the SVJ model using the
Kyriakou-Brignone-Fusai (2024) method for the simulation.
}
\examples{
S = 100; K = 100; v0 = 0.008836; k = 3.99; theta = 0.014; sigma = 0.27
rho = -0.79; r = 0.0319; tau = 5; lambda = 0.11; mu_bar = -0.12
sigma_s = 0.15; true_price = 20.1642
# price_svj(10000, S, K, v0, tau, r, k, theta, sigma, rho, lambda, mu_bar,
# sigma_s, true_price)
}
