% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryield_svcj.R
\name{ryield_svcj}
\alias{ryield_svcj}
\title{Simulate Yield using Kyriakou-Brignone-Fusai method for the SVCJ model}
\usage{
ryield_svcj(
  n,
  v0,
  tau,
  r,
  k,
  theta,
  sigma,
  rho,
  lambda,
  mu_bar,
  sigma_s,
  mu_v,
  rho_J
)
}
\arguments{
\item{n}{number of yield samples to simulate.}

\item{v0}{current variance level.}

\item{tau}{time difference, \eqn{t-u}.}

\item{r}{riskless rate, risk-neutral drift.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}

\item{rho}{parameter \eqn{\rho}.}

\item{lambda}{parameter \eqn{\lambda}.}

\item{mu_bar}{parameter \eqn{\bar{\mu}}.}

\item{sigma_s}{parameter \eqn{\sigma_s}.}

\item{mu_v}{parameter \eqn{\mu_v}, mean of the jumps
(exponential distribution) in the variance.}

\item{rho_J}{parameter \eqn{\rho_J}, correlation parameter between the
jumps in the price and variance.}
}
\value{
next period yield, a scalar.
}
\description{
Simulate the next period yield given current variance using the
Kyriakou-Brignone-Fusai (2024) method for the SVCJ model.
}
\examples{
v0 = 0.007569; k = 3.46; theta = 0.008; sigma = 0.14; rho = -0.82
r = 0.0319; tau = 1; lambda = 0.47; mu_bar = -0.1; sigma_s = 0.0001
mu_v = 0.05; rho_J = -0.38
# Y = ryield_svcj(1000, v0, tau, r, k, theta, sigma, rho,
# lambda, mu_bar, sigma_s, mu_v, rho_J)
# hist(Y)
}
