% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riv.R
\name{riv}
\alias{riv}
\title{Simulating IV from Conditional Distribution}
\usage{
riv(v0, v1, tau, k, theta, sigma)
}
\arguments{
\item{v0}{the left end, \eqn{v_u}.}

\item{v1}{the right end, \eqn{v_t}.}

\item{tau}{time difference, \eqn{t-u}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}
}
\value{
scalar
}
\description{
Random variable generation for the conditional distribution of the
Integrated Variance (IV), through a Pearson Distribution approximation
which match the first four moments of the unknown true conditional
distribution of IV.
}
\details{
The first four moments are calculated through \code{\link[=invMGF]{invMGF()}}.
\strong{Note that:}
\itemize{
\item when the time difference \code{tau} is very short, like < 0.0001, it is very
likely to introduce computational errors, therefore, we let
\code{iv = tau * (v0+v1)/2} directly.
\item when the computed variance is very small, the computed skewness and
kurtosis are highly un-reliable, therefore, we let \code{iv} equal the mean
directly.
}
}
\examples{
k = 6.21; theta = 0.019; sigma = 0.61; v0 = 0.010201; v1 = v0; tau = 1
iv = riv(v0, v1, tau, k, theta, sigma)
}
\references{
\enumerate{
\item Kyriakou, I., Brignone, R., & Fusai, G. (2024). Unified moment-based modeling
of integrated stochastic processes. \emph{Operations Research}, 72(4), 1630-1653.
\item Broadie, M., & Kaya, Ö. (2006). Exact simulation of stochastic volatility
and other affine jump diffusion processes. \emph{Operations Research},
54(2), 217-231.
}
}
\seealso{
See \code{\link[ajd.sim.bk:piv]{ajd.sim.bk::riv()}} for the Broadie-Kaya algorithm which
is through Fourier inversion of its Characteristic Function.
}
