% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryield_Hest.R
\name{ryield_Hest}
\alias{ryield_Hest}
\title{Simulate the Yield using Kyriakou-Brignone-Fusai Method}
\usage{
ryield_Hest(n, v0, tau, r, k, theta, sigma, rho)
}
\arguments{
\item{n}{number of yield samples to simulate.}

\item{v0}{current variance level.}

\item{tau}{time difference, \eqn{t-u}.}

\item{r}{riskless rate, risk-neutral drift.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma}{parameter \eqn{\sigma}.}

\item{rho}{parameter \eqn{rho}.}
}
\value{
next period yield, a scalar.
}
\description{
Simulate the next period yield given current variance using the
Kyriakou-Brignone-Fusai (2024) method.
}
\examples{
v0 = 0.010201; k = 6.21; theta = 0.019; sigma = 0.61; rho = -0.7
r = 0.0319; tau = 1
Y = ryield_Hest(10, v0, tau, r, k, theta, sigma, rho)
hist(Y)
}
